AMDA object is now shared through pyramid registry, and fixed issue on server kill

+++ sciqlopcache/test_cache.py
@@ -148,7 +148,8 @@ class _CacheTest(unittest.TestCase):
 
     def tearDown(self):
         del self.cache
-        os.remove(self.dbfile)
+        if os.path.exists(self.dbfile):
+            os.remove(self.dbfile)
 
 
 @ddt
